Added GiochiPossedutiController and fixed classes DAO JDBC classes

+++ src/main/java/it/unical/mat/webcomp21/persistence/dao/GiocoPossedutoDAO.java
@@ -7,7 +7,7 @@ import it.unical.mat.webcomp21.model.GiocoPosseduto;
 public interface GiocoPossedutoDAO {
 	
 	public void save(GiocoPosseduto gd);  // Create
-	public GiocoPosseduto findByPrimaryKey(int id);     // Retrieve
+	public GiocoPosseduto findByPrimaryKey(long id);     // Retrieve
 	public List<GiocoPosseduto> findAll();       
 	public void update(GiocoPosseduto gd); //Update
 	public void delete(GiocoPosseduto gd); //Delete
